+++ bin/run_plate.py
@@ -17,28 +17,43 @@ def main(arguments):
     parser.add_argument('plate', help="plate label, ie miseq-plate-{label}")
     parser.add_argument('-d', '--data-dir',
                         default='/fh/fast/fredricks_d/bvdiversity/data')
+    parser.add_argument('-o', '--out-dir', default='dada2_nf_out')
     parser.add_argument('-f', '--force', action='store_true', default=False,
-                        help='do not prompt for confirmation before launching pipleine')
+                        help="""do not prompt for confirmation before
+                        launching pipleine""")
     parser.add_argument('-c', '--check-inputs', action='store_true', default=False,
                         help='verify inputs and exit')
+    parser.add_argument('-p', '--profile', choices=['hutch_batch', 'singularity'],
+                        default='hutch_batch', help="""[%(default)s]""")
 
     args = parser.parse_args(arguments)
     plate = args.plate
 
     platedir = path.join(args.data_dir, f'miseq-plate-{plate}')
-    outdir = path.join(args.data_dir, 'dada2_nf_out', f'miseq-plate-{plate}')
+    outdir = path.join(args.data_dir, args.out_dir, f'miseq-plate-{plate}')
     try:
         os.makedirs(outdir)
     except OSError as err:
         pass
 
-    # fastq list
-    fq_pattern = path.join(
-        platedir,
-        f'run-files/*/Data/Intensities/BaseCalls/m{plate}*.fastq.gz')
-    files = glob(fq_pattern)
+    # fastq list - the directory layout has changed over time, so we
+    # need to try a few patterns
+    layouts = [
+        f'run-files/*/Data/Intensities/BaseCalls/m{plate}*.fastq.gz',
+        f'run-files/*/Alignment_1/*/Fastq/m{plate}*.fastq.gz',  # most recent format
+        f'run-files/*/*/m{plate}*.fastq.gz',  # eg, plate 3
+    ]
 
-    print('found {} fastq files'.format(len(files)))
+    for pattern in layouts:
+        fq_pattern = path.join(platedir, pattern)
+        print(f'searching {fq_pattern}')
+        files = glob(fq_pattern)
+        print('found {} fastq files'.format(len(files)))
+        if files:
+            break
+    else:
+        print('no files found, exiting')
+        sys.exit(1)
 
     fastq_list = path.abspath(path.join(outdir, 'fastq_list.txt'))
     with open(fastq_list, 'w') as f:
@@ -56,6 +71,7 @@ def main(arguments):
         'output': path.join(outdir, 'output'),
         'index_file_type': 'dual',
         'dada_params': 'data/dada_params_300.json',
+        'alignment_model': 'data/SSU_rRNA_bacteria.cm',
     }
     with open(params_file, 'w') as f:
         json.dump(d, f, indent=2)
@@ -73,7 +89,9 @@ def main(arguments):
     for cmd in cmds:
         subprocess.run(cmd, shell=True)
 
-    runcmd = f'nextflow run main.nf -profile hutch_batch -params-file {params_file}'
+    runcmd = (f'nextflow run main.nf '
+              f'-profile {args.profile} '
+              f'-params-file {params_file}')
     print(runcmd)
     response = 'yes' if args.force else input('Run this pipleine? (yes/no) ')
     if response == 'yes':